Fixes on index.php logic

+++ index.php
@@ -16,117 +16,92 @@
 include_once(getcwd() . "/vendor/autoload.php");
 use KubAT\PhpSimple\HtmlDomParser;
 
-class EconomicCalendarAPI
-{
+class EconomicCalendarAPI {
     private $cacheSuffix = "_api-cache.json";
-    private $minGapFromNextEvent = 300; // Cache is disabled 5 min after/before an event
+    private $minGapFromNextEvent = 300; // If next event is timed for the next 5 minutes, or if it was less than 5 minutes ago, cache won't be used
     private $dataSourceUrl = "https://sslecal2.forexprostools.com/";
     private $cacheFile = false; // File path to cache
 
-    public function fetchEvents()
-    {
+    public function fetchEvents() {
         header("Content-Type: application/json");
-
+        
         $this->setupCacheFile();
 
         if ($this->shouldUseCache()) {
-            header("API-Info-Using-Cache: true");
             $this->serveFromCache();
         } else {
-            header("API-Info-Using-Cache: false");
             $this->fetchAndCacheData();
         }
     }
-
-    private function setupCacheFile()
-    {
+    
+    private function setupCacheFile() {
         $this->cacheFile = $this->findFirstCacheFile();
     }
 
-    private function shouldUseCache()
-    {
-        if (!$this->cacheFile)
-            return false;
+    private function shouldUseCache() {
+        if (!$this->cacheFile) return false;
 
         $cacheTime = $this->extractTimestampFromFilePath($this->cacheFile);
-        return ($cacheTime - time()) > $this->minGapFromNextEvent;
+        return ($cacheTime - time()) > $this->minGapFromNextEvent && (time() - $cacheTime) < $this->minGapFromNextEvent;
     }
 
-    private function shouldCache()
-    {
-        $cacheTime = $this->extractTimestampFromFilePath($this->cacheFile);
-        return (time() - $cacheTime) > $this->minGapFromNextEvent; // Just caches after minutes of gap are passed
-    }
-
-
-    private function serveFromCache()
-    {
+    private function serveFromCache() {
         echo file_get_contents($this->cacheFile);
         exit();
     }
 
-    private function fetchAndCacheData()
-    {
+    private function fetchAndCacheData() {
         $dom = HtmlDomParser::file_get_html($this->dataSourceUrl);
         $data = $this->parseDataFromDom($dom);
-
-        if ($this->shouldCache()) {
-            $this->cacheData($data);
-        }
-
-        echo json_encode($data['events']);
+        $this->cacheData($data);
+        echo json_encode($data)["events"];
         exit();
     }
 
-    private function findFirstCacheFile()
-    {
+    private function findFirstCacheFile() {
         $files = glob("./*" . $this->cacheSuffix);
         return $files ? $files[0] : null;
     }
 
-    private function extractTimestampFromFilePath($filePath)
-    {
+    private function extractTimestampFromFilePath($filePath) {
         if (preg_match('/(\d+)' . preg_quote($this->cacheSuffix) . '$/', $filePath, $matches)) {
-            return (int) $matches[1];
+            return (int)$matches[1];
         }
-        return 0;
+        return null;
     }
 
-    private function getMinuteDifference($dateString)
-    {
+    private function getMinuteDifference($dateString) {
         $specifiedDate = new DateTime($dateString);
         $serverDate = new DateTime("now");
         return ($specifiedDate->getTimestamp() - $serverDate->getTimestamp()) / 60;
     }
 
-    private function sanitize($str)
-    {
+    private function sanitize($str) {
         return trim(str_replace("&nbsp;", "", $str));
     }
 
-    private function parseDataFromDom($dom)
-    {
+    private function parseDataFromDom($dom) {
         $elems = $dom->getElementById("#ecEventsTable")->find("tr[id*='eventRowId']");
         $events = [];
 
         $nextFutureEvent = null;
         $nextFutureEventSmallestDif = 9999;
-
-        foreach ($elems as $element) {
+                
+        foreach($elems as $element) {
             $date = isset($element->attr["event_timestamp"]) ? $element->attr["event_timestamp"] : null;
             $name = count($name = $element->find("td.event")) > 0 ? $this->sanitize($name[0]->text()) : null;
 
-            if ($date != null) {
+            if($date != null) {
                 $minutesDifference = $this->getMinuteDifference($date);
-                if ($minutesDifference > 0 && ($nextFutureEvent == null || $minutesDifference < $nextFutureEventSmallestDif)) {
+                if($minutesDifference > 0 && ($nextFutureEvent == null || $minutesDifference < $nextFutureEventSmallestDif)) {
                     $nextFutureEvent = $date;
                     $nextFutureEventSmallestDif = $minutesDifference;
                 }
             }
-
+            
             array_push($events, [
                 "economy" => $this->sanitize($element->find("td.flagCur")[0]->text()),
-                "impact" => count($element->find("td.sentiment")[0]->find("i.grayFullBullishIcon")),
+                "impact" =>  count($element->find("td.sentiment")[0]->find("i.grayFullBullishIcon")),
                 "date" => $date,
                 "name" => $name,
                 "actual" => count($actual = $element->find("td.act")) > 0 ? $this->sanitize($actual[0]->text()) : null,
@@ -136,16 +111,14 @@ class EconomicCalendarAPI
         }
 
         return [
-            "next_event_date" => $nextFutureEvent,
-            "events" => $events,
+            "next_event_date"=> $nextFutureEvent,
+            "events"=> $events,
         ];
     }
 
-    private function cacheData($data)
-    {
-        if ($this->cacheFile)
-            unlink($this->cacheFile);
-
+    private function cacheData($data) {
+        if ($this->cacheFile) unlink($this->cacheFile);
+        
         $nextEventDate = new DateTime($data['next_event_date']);
         $cacheFileName = $nextEventDate->getTimestamp() . $this->cacheSuffix;
         file_put_contents($cacheFileName, json_encode($data['events']));
